Wolf AI Change
So, turns out using that many layermasks is irresponsible, because of the limited number available, the fact that having to iterate over them slows down the system when it could do it once and then sort, and I got tired of using so many different layermasks
Instead, only one layer, and uses tags to sort them after finding all the junctions.
In addition, the NameToLayer is kinda going weird so i'm using bit shifting to make the layermask inversions instead.