Wolf Maze level and functionality
Made the main structure of the maze level with the wolf, and all the Wolf-side scripts.

Wolf raycasting now works. He waits for the player to be in line, then raycasts for Wall/Player, and if it's a player then runs. He raycasts in each other direction for Wall/Junction, and if it's a junction tracks it. Then, it chooses the highest prority junction. It marks down the target location, goes to it, then resets.
If the junction was a dead end, it stops thinking. Eventually, it will need to drop the leg and escape somehow.
The drop-able walls will work as well, just need to program in a new Interactable and animate something

Next functions: Wolf turning in the direction it runs, dropping the leg off when trapped, the levers/walls