WolfDropAnimation and XOR Room
XOR Room was easy. Made a pressureplate, and gave it a second box collider. Never would have worked so easily if I didn't make it so robust earlier, so thank you, past Matt. Worked out.

WolfDropAnimation is here as well. Wolf puts down the leg and runs away, while calling a couple of functions to make the leg object appear, as well as the wolf disappear.

However, the wolf is now not moving. This is a temporary commit so I can roll back and compare the problems.

Hopefully this goes well.